backend/requirements.txt
asgiref==3.6.0
asttokens==2.2.1
attrs==23.1.0
autoprop==4.1.0
backcall==0.2.0
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.0.1
colorama==0.4.6
colorlog==6.7.0
comm==0.1.3
cryptography==39.0.1
Cython==0.29.34
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dj-rest-auth==3.0.0
Django==4.1.7
django-allauth==0.52.0
django-cors-headers==3.13.0
django-extensions==3.2.3
djangorestframework==3.14.0
executing==1.2.0
glooey==0.3.6
idna==3.4
ifcopenshell==0.7.0.230418
imageio==2.28.1
ipykernel==6.23.3
ipython==8.14.0
jedi==0.18.2
jsonschema==4.17.3
jupyter_client==8.3.0
jupyter_core==5.3.1
lark==1.1.5
lazy_loader==0.2
lxml==4.9.2
mapbox-earcut==1.0.1
markdown-it-py==2.2.0
matplotlib-inline==0.1.6
mdurl==0.1.2
meshio==5.3.4
more-itertools==9.1.0
mpmath==1.3.0
nest-asyncio==1.5.6
networkx==3.1
numpy==1.24.3
oauthlib==3.2.2
packaging==23.1
pandas==2.0.2
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.8.0
prompt-toolkit==3.0.38
psutil==5.9.5
pure-eval==0.2.2
pycollada==0.7.2
pycparser==2.21
pydot==1.4.2
pyglet==1.5.27
Pygments==2.15.1
PyJWT==2.6.0
pyparsing==3.1.0
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
pyrsistent==0.19.3
PySide6==6.5.1.1
PySide6-Addons==6.5.1.1
PySide6-Essentials==6.5.1.1
python-dateutil==2.8.2
python-fcl==0.7.0.4
python3-openid==3.2.0
pytz==2023.3
PyWavelets==1.4.1
pywin32==306
PyYAML==6.0
pyzmq==25.1.0
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.3.5
Rtree==1.0.1
scikit-image==0.20.0
scipy==1.10.1
shapely==2.0.1
shiboken6==6.5.1.1
signature_dispatch==1.0.1
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
svg.path==6.3
sympy==1.12
tifffile==2023.4.12
tk==0.1.0
tornado==6.3.2
traitlets==5.9.0
trimesh==3.21.6
typeguard==3.0.2
tzdata==2023.3
urllib3==1.26.14
vecrec==0.3.1
wcwidth==0.2.6
xatlas==0.0.7
xxhash==3.2.0
